# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Neutrino API",
    "version": "1.0.0"
  },
  "paths": {
    "/neutrino/process": {
      "post": {
        "operationId": "post_heavy_processing",
        "summary": "Heavy Processing",
        "tags": [],
        "description": "\n    CPU-intensive task that benefits from Neutrino's worker pool.\n\n    This would be slow in a standard async FastAPI setup because it\n    blocks the event loop. With Neutrino, it runs in a separate worker\n    process with proper resource management.\n    ",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "text": {
                    "title": "Text",
                    "type": "string"
                  },
                  "iterations": {
                    "title": "Iterations",
                    "type": "integer"
                  }
                },
                "required": [
                  "text",
                  "iterations"
                ],
                "title": "TaskRequest",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "result": {
                      "title": "Result",
                      "type": "string"
                    },
                    "processed_chars": {
                      "title": "Processed Chars",
                      "type": "integer"
                    }
                  },
                  "required": [
                    "result",
                    "processed_chars"
                  ],
                  "title": "TaskResponse",
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/neutrino/analyze": {
      "post": {
        "operationId": "post_analyze_data",
        "summary": "Analyze Data",
        "tags": [],
        "description": "\n    Statistical analysis task.\n\n    In a real application, this might involve ML inference, image processing,\n    or other compute-heavy operations that benefit from Neutrino's\n    orchestration and worker lifecycle management.\n    ",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "user_id": {
                    "title": "User Id",
                    "type": "integer"
                  },
                  "data": {
                    "items": {
                      "type": "number"
                    },
                    "title": "Data",
                    "type": "array"
                  }
                },
                "required": [
                  "user_id",
                  "data"
                ],
                "title": "AnalysisRequest",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user_id": {
                      "title": "User Id",
                      "type": "integer"
                    },
                    "mean": {
                      "title": "Mean",
                      "type": "number"
                    },
                    "median": {
                      "title": "Median",
                      "type": "number"
                    },
                    "std_dev": {
                      "title": "Std Dev",
                      "type": "number"
                    }
                  },
                  "required": [
                    "user_id",
                    "mean",
                    "median",
                    "std_dev"
                  ],
                  "title": "AnalysisResponse",
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TaskRequest": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          },
          "iterations": {
            "title": "Iterations",
            "type": "integer"
          }
        },
        "required": [
          "text",
          "iterations"
        ],
        "title": "TaskRequest",
        "type": "object"
      },
      "TaskResponse": {
        "properties": {
          "result": {
            "title": "Result",
            "type": "string"
          },
          "processed_chars": {
            "title": "Processed Chars",
            "type": "integer"
          }
        },
        "required": [
          "result",
          "processed_chars"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
      "AnalysisRequest": {
        "properties": {
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "data": {
            "items": {
              "type": "number"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "user_id",
          "data"
        ],
        "title": "AnalysisRequest",
        "type": "object"
      },
      "AnalysisResponse": {
        "properties": {
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "mean": {
            "title": "Mean",
            "type": "number"
          },
          "median": {
            "title": "Median",
            "type": "number"
          },
          "std_dev": {
            "title": "Std Dev",
            "type": "number"
          }
        },
        "required": [
          "user_id",
          "mean",
          "median",
          "std_dev"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      }
    }
  },
  "x-neutrino-asgi": {
    "enabled": true,
    "app_class": "fastapi.applications.FastAPI"
  }
}